refactor: context.jsx 파일 분리 및 App단으로 변경
- min27604/baseball/#10
