+++ FE/react-baseball/src/components/startScreen/Caption.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useContext } from 'react';
-import { PlayableContext } from '../../routes/StartScreen';
+import { PlayableContext } from '../provider/Context';
 
 const Caption = (props) => {
   const playable = useContext(PlayableContext);
